Add files via upload
Example script to use real-time RAP analysis fields -- accessed via Unidata's THREDDS server -- to compute and display 1000-500-hPa mean virtual temperature and geopotential thickness to visualize terms in the hypsometric equation. This script produces maps for the eastern U.S. and was set up for a wintertime event. Provided as is.